feat: implement context in the app Todo.
