+++ src/components/types.tsx
@@ -9,3 +9,15 @@ export enum FilterCases {
   COMPLETE,
   ACTIVE,
 }
+
+export interface ITodoContext {
+  taskList: Task[];
+  selectedTask: Task | undefined;
+  taskText: string;
+  addTask: (textTask: string) => void;
+  startEditing: (id: string) => void;
+  editTask: () => void;
+  deleteTask: (id: string) => void;
+  checkTask: (id: string) => void;
+  setTaskText: (id: string) => void;
+}
